add docstrings and improve modular structure
